dl/trm_crypto_signal_project/requirements.txt
ccxt 
pandas 
numpy 
pandas_ta 
python-binance 
python-telegram-bot 
streamlit 
ta
